move CreateSessionAttributes to more appropriate types folder, move relevant session type as well
